more compile error fixed

+++ src/GPU/VariablesCUDA.cuh
@@ -105,6 +105,6 @@ public:
   double *gpu_aForcex, *gpu_aForcey, *gpu_aForcez;
   double *gpu_mForcex, *gpu_mForcey, *gpu_mForcez;
   double *gpu_rMin, *gpu_expConst, *gpu_rMaxSq;
-  double *gpu_cellVector;
+  int *gpu_cellVector;
 };
 #endif
